reafctor: 채팅 내역 조회 API 연결
