+++ app/(hosts)/chatlist/[chatlistId]/page.jsx
@@ -3,12 +3,14 @@ import {SkeletonLoader} from "@/app/ui/skeletons";
 import {CardWrapper} from "@/app/ui/(host)/chatlist/Cards";
 import {ChatListHeader} from "@/app/ui/(host)/chatlist/Header";
 
-export default function Page() {
+export default async function Page({ params }) {
+    const {chatlistId} = await params;
+
     return (
         <>
             <ChatListHeader/>
             <Suspense fallback={<SkeletonLoader/>}>
-                <CardWrapper/>
+                <CardWrapper chatroomId={chatlistId}/>
             </Suspense>
         </>
     );
