+++ app/ui/Card.jsx
@@ -2,11 +2,11 @@ import DropdownMenuDemo from "@/app/ui/(host)/chatlist/Dropdown";
 import LikeButton from "@/app/ui/(host)/chatlist/Likes";
 
 export default function Card({card}) {
-    const {id,is_pinned, content, username, likes} = card;
+    const {id, pinned, content, username, likes} = card;
 
     return (
         <div className={`p-4 rounded-lg shadow mb-4 ${
-            is_pinned ? 'bg-yellow-100' : 'bg-white'
+            pinned ? 'bg-yellow-100' : 'bg-white'
         }`}>
             <div className={"flex justify-between"}>
